Corrigido classe Bench

+++ Assignments/app/src/main/kotlin/org/example/classes/mainClasses.kt
@@ -44,7 +44,7 @@ data class CasoDetalhado(
 
 data class EstatisticaAbuso(
     @Column("EAbusiva") val abusiva: Boolean,
-    @Column("Total") val total: Int,
+    @Column("Total") val total: Long,
 )
 
 data class CasosPorInteracao(
